+++ Apizr/Src/Caching/Apizr.Extensions.Microsoft.Caching/MicrosoftCachingOptionsBuilderExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Apizr.Extending.Configuring.Common;
+﻿using Apizr.Configuring.Common;
+using Apizr.Extending.Configuring.Common;
 
 [assembly: Apizr.Preserve]
 namespace Apizr
@@ -12,9 +13,9 @@ namespace Apizr
         /// Use any registered IDistributedCache implementation
         /// </summary>
         /// <returns></returns>
-        public static TBuilder WithDistributedCacheHandler<TBuilder, TCache>(this TBuilder builder) where TBuilder : IApizrExtendedCommonOptionsBuilder
+        public static TBuilder WithDistributedCacheHandler<TBuilder, TCache>(this TBuilder builder) where TBuilder : IApizrExtendedCommonOptionsBuilderBase
         {
-            builder.WithCacheHandler<DistributedCacheHandler<TCache>>();
+            builder.SetCacheHandlerType<DistributedCacheHandler<TCache>>();
 
             return builder;
         }
@@ -24,9 +25,9 @@ namespace Apizr
         /// Use any registered IMemoryCache implementation
         /// </summary>
         /// <returns></returns>
-        public static TBuilder WithInMemoryCacheHandler<TBuilder>(this TBuilder builder) where TBuilder : IApizrExtendedCommonOptionsBuilder
+        public static TBuilder WithInMemoryCacheHandler<TBuilder>(this TBuilder builder) where TBuilder : IApizrExtendedCommonOptionsBuilderBase
         {
-            builder.WithCacheHandler<InMemoryCacheHandler>();
+            builder.SetCacheHandlerType<InMemoryCacheHandler>();
 
             return builder;
         }